Extract callbacks to a concern
